requirements.txt
numpy
matplotlib
pillow
torch
torchvision
opencv-python
transformers
seaborn
scipy
opencv-python
#tqdm
#pandas
#nltk
#multimatch_gaze
#fastdtw
#scikit-klearn